MBS-9069 move MockDispatcherTest
