+++ subprojects/android-test/test-inhouse-runner/src/test/kotlin/com/avito/android/monitoring/StatsdHttpTrackerKtTest.kt
@@ -1,7 +1,6 @@
 package com.avito.android.monitoring
 
 import com.google.common.truth.Truth.assertThat
-import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
 import org.junit.Test
